requirements.txt
asgiref==3.9.1
Django==4.2.23
djangorestframework==3.16.0
pillow==11.3.0
sqlparse==0.5.3
typing_extensions==4.14.1
djangorestframework-simplejwt==5.5.1
celery==5.4.0
redis==6.3.0
ffmpeg-python==0.2.0
eventlet==0.40.2 
channels==4.0.0
channels_redis==4.2.0
uvicorn[standard]==0.35.0
cryptography==45.0.6
django-cors-headers==4.6.0
